导出onnx int32

+++ task_pretrain/poetry_t5_pretrain/task_poetry_t5.py
@@ -152,10 +152,10 @@ if __name__ == '__main__':
                                                        model_args=model_args,
                                                        training_args=training_args)
         input_sample = (
-            ("input_ids", torch.ones(size=(1, 128), dtype=torch.int64)),
-            ("attention_mask", torch.ones(size=(1, 128), dtype=torch.int64)),
-            ("decoder_input_ids", torch.ones(size=(1, 128), dtype=torch.int64)),
-            ("decoder_attention_mask", torch.ones(size=(1, 128), dtype=torch.int64)),
+            ("input_ids", torch.ones(size=(1, 128), dtype=torch.int32)),
+            ("attention_mask", torch.ones(size=(1, 128), dtype=torch.int32)),
+            ("decoder_input_ids", torch.ones(size=(1, 128), dtype=torch.int32)),
+            ("decoder_attention_mask", torch.ones(size=(1, 128), dtype=torch.int32)),
         )
         input_names = ("input_ids","attention_mask", "decoder_input_ids","decoder_attention_mask")
         output_names = ("pred_ids",)
